Feature: Implement a presolver for GF(2) equations.
