+++ printemps/model_component/constraint.h
@@ -562,17 +562,19 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
 
             if (is_valid) {
                 for (const auto &sensitivity : m_expression.sensitivities()) {
-                    if (!is_integer(sensitivity.second)) {
+                    auto variable_ptr = sensitivity.first;
+                    auto coefficient  = sensitivity.second;
+
+                    if (!is_integer(coefficient)) {
                         is_valid = false;
                         break;
                     }
-                    if ((sensitivity.first->sense() == VariableSense::Integer ||
-                         sensitivity.first->sense() ==
+                    if ((variable_ptr->sense() == VariableSense::Integer ||
+                         variable_ptr->sense() ==
                              VariableSense::Intermediate) &&
-                        abs(sensitivity.second) == 1 &&
-                        !sensitivity.first->is_fixed()) {
-                        aux_variable_ptr         = sensitivity.first;
-                        aux_variable_coefficient = sensitivity.second;
+                        abs(coefficient) == 1 && !variable_ptr->is_fixed()) {
+                        aux_variable_ptr         = variable_ptr;
+                        aux_variable_coefficient = coefficient;
                         number_of_integer_variables_with_coefficient_one++;
                     }
                 }
@@ -655,16 +657,22 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
 
             if (is_valid) {
                 for (const auto &sensitivity : m_expression.sensitivities()) {
-                    if ((sensitivity.first->sense() == VariableSense::Integer ||
-                         sensitivity.first->sense() == VariableSense::Binary) &&
-                        abs(sensitivity.second) == 2 &&
-                        !sensitivity.first->is_fixed()) {
-                        aux_variable_ptr         = sensitivity.first;
-                        aux_variable_coefficient = sensitivity.second;
+                    auto variable_ptr = sensitivity.first;
+                    auto coefficient  = sensitivity.second;
+
+                    if (variable_ptr->is_fixed()) {
+                        is_valid = false;
+                        break;
+                    }
+
+                    if ((variable_ptr->sense() == VariableSense::Integer ||
+                         variable_ptr->sense() == VariableSense::Binary) &&
+                        abs(coefficient) == 2) {
+                        aux_variable_ptr         = variable_ptr;
+                        aux_variable_coefficient = coefficient;
                         number_of_integer_variables_with_coefficient_two++;
-                    } else if (sensitivity.first->sense() !=
-                                   VariableSense::Binary ||
-                               abs(sensitivity.second) != 1) {
+                    } else if (variable_ptr->sense() != VariableSense::Binary ||
+                               abs(coefficient) != 1) {
                         is_valid = false;
                         break;
                     }
@@ -726,8 +734,8 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
 
     /*************************************************************************/
     inline constexpr T_Expression evaluate_constraint(
-        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE) const
-        noexcept {
+        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE)
+        const noexcept {
 #ifdef _MPS_SOLVER
         return m_expression.evaluate(a_MOVE);
 #else
@@ -749,15 +757,15 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
 
     /*************************************************************************/
     inline constexpr T_Expression evaluate_violation(
-        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE) const
-        noexcept {
+        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE)
+        const noexcept {
         return m_violation_function(a_MOVE);
     }
 
     /*************************************************************************/
     inline constexpr T_Expression evaluate_violation_diff(
-        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE) const
-        noexcept {
+        const neighborhood::Move<T_Variable, T_Expression> &a_MOVE)
+        const noexcept {
         return m_violation_function(a_MOVE) - m_violation_value;
     }
 
@@ -847,8 +855,8 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
     }
 
     /*************************************************************************/
-    inline constexpr double local_penalty_coefficient_less(void) const
-        noexcept {
+    inline constexpr double local_penalty_coefficient_less(
+        void) const noexcept {
         return m_local_penalty_coefficient_less;
     }
 
@@ -857,8 +865,8 @@ class Constraint : public multi_array::AbstractMultiArrayElement {
         return m_local_penalty_coefficient_greater;
     }
     /*************************************************************************/
-    inline constexpr double local_penalty_coefficient_greater(void) const
-        noexcept {
+    inline constexpr double local_penalty_coefficient_greater(
+        void) const noexcept {
         return m_local_penalty_coefficient_greater;
     }
