+++ printemps/solver/solver.h
@@ -6,12 +6,14 @@
 #ifndef PRINTEMPS_SOLVER_H__
 #define PRINTEMPS_SOLVER_H__
 
-#include "tabu_search_trend_logger.h"
+#include "memory.h"
 #include "status.h"
 #include "result.h"
-#include "tabu_search/tabu_search.h"
-#include "local_search/local_search.h"
+#include "tabu_search_trend_logger.h"
+
 #include "lagrange_dual/lagrange_dual.h"
+#include "local_search/local_search.h"
+#include "tabu_search/tabu_search.h"
 
 namespace printemps {
 namespace solver {
